==> R/2-modeling/1-Compute_NNI.Rmd <==
---
title: "Compute NNI of the intercrops"
classoption: a4paper
output:
  pdf_document:
    fig_caption: no
    fig_crop: no
urlcolor: blue
---


```{r setup, warning=FALSE, message=FALSE, echo=FALSE}
# packages
library(data.intercrop)
library(readxl)
library(lubridate)
library(knitr)
library(readr)
library(dplyr)
library(tidyr)
library(ggplot2)
# options
theme_set(theme_bw())
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
opts_chunk$set(echo=FALSE,
               warning=FALSE,
               message=FALSE,
               fig.width = 8,
               fig.height = 7,
               dpi = 200)
options(knitr.kable.NA = '')
theme_set(theme_bw())

traits_mean <- data.intercrop::traits_mean
traits_crop <- data.intercrop::traits$crop
index <- data.intercrop::index


experiments_paper <- data.intercrop::index %>% filter(experiment_id %in% 
c("Auzeville_SGs_2007", "Auzeville_TE_2006", "Auzeville_TO_2013",  "Auzeville_cochard_2010", "Auzeville_PP_2011", "Auzeville_ZN_2012") ) %>% distinct(experiment_id)

index_paper <- data.intercrop::index %>% inner_join(experiments_paper)



```

```{r define functions}


which.customized <- function (x, arr.ind = FALSE, useNames = TRUE) 
{
    result = which(x, arr.ind = FALSE, useNames = TRUE)
    ifelse(identical(result, integer(0)), return(NA), return(result))
}
```

# Compute NNI

```{r}


#Select relevant variables and specify species' parameters
data_compute_NNI <- bind_rows(traits_mean, traits_crop) %>% 
  inner_join(experiments_paper) %>% 
  filter(species_mix %in% c("wheat_turgidum", "pea", "fababean", "wheat_turgidum_pea", "wheat_turgidum_fababean")) %>% 
  filter(variable %in% c("nitrogen_abs_shoot", "nitrogen_shoot", "biomass_shoot", "height")) %>% 
  select(experiment_id:value, season) %>% 
  mutate(ac_sp = case_when(
      species %in% c("pea","fababean") ~ 5.1,
    stringr::str_detect(species, "wheat") & season == "winter" ~ 5.3
  )) %>% 
  mutate(bc_sp = case_when(
      species %in% c("pea","fababean") ~ 0.32,
      stringr::str_detect(species, "wheat") & season == "winter" ~ 0.44
    )) %>% 
  # filter(crop_type =="IC") %>%
  select(-c(sowing_date, harvest_date, density_relative)) %>% 
  distinct


# For each date, compute total biomass by IC and relative biomass (parameter p_spi)
data_biomass_shoot <- data_compute_NNI%>% 
  filter(variable == "biomass_shoot") %>% 
  group_by(experiment_id, variable, management,species_mix, date ) %>% 
  mutate(biomass_shoot_total = if_else(crop_type == "IC",value[which.customized(plant_family == "cereal")]+
           value[which.customized(plant_family == "legume")], value),
           biomass_shoot_component = value) %>% 
  mutate(p_spi = biomass_shoot_component/biomass_shoot_total) %>% 
  ungroup %>% 
  select(experiment_id, season, crop_type, management,species_mix,plant_family,
         time_thermal,
         species,  date , contains("biomass"), p_spi) 



data_nitrogen_shoot <- data_compute_NNI%>% 
  filter(variable == "nitrogen_shoot") %>% 
  mutate(nitrogen_shoot_component  = value) %>% 
  select(experiment_id,season,  crop_type,  management,species_mix,plant_family, species,  date ,
         time_thermal, contains("nitrogen")) 

data_nitrogen_abs_shoot <- data_compute_NNI%>% 
  filter(variable == "nitrogen_abs_shoot") %>% 
  group_by(experiment_id, variable, management,species_mix, date ) %>% 
  mutate(nitrogen_abs_shoot_IC = value[which.customized(plant_family == "cereal")]+
           value[which.customized(plant_family == "legume")],
           nitrogen_abs_shoot_component = value) %>% 
  ungroup %>% 
  select(experiment_id,season, crop_type, management,species_mix,plant_family, species,  date ,
         time_thermal, contains("nitrogen_abs")) 


# See Table 1 Louarn2019 for parameters names
data_compute_NNI_IC_SC<- full_join(data_biomass_shoot, data_nitrogen_shoot) %>% 
  full_join(data_nitrogen_abs_shoot) %>% 
  full_join(data_compute_NNI %>% distinct(species, season,  ac_sp,bc_sp)) %>% 
  group_by(experiment_id, management,species_mix, date) %>% 
  mutate(ac_sp_cereal = ac_sp[which.customized(plant_family == "cereal")],
         ac_sp_legume = ac_sp[which.customized(plant_family == "legume")],
         bc_sp_cereal = bc_sp[which.customized(plant_family == "cereal")],
         bc_sp_legume = bc_sp[which.customized(plant_family == "legume")]
         ) %>% 
  ungroup %>% 
  mutate(ac_sp_i = if_else(plant_family == "cereal", ac_sp_cereal, ac_sp_legume),
         ac_sp_j = if_else(plant_family == "cereal", ac_sp_legume, ac_sp_cereal),
         bc_sp_i = if_else(plant_family == "cereal", bc_sp_cereal, bc_sp_legume),  
         bc_sp_j = if_else(plant_family == "cereal", bc_sp_legume, bc_sp_cereal) ) %>% 
  select(-c(ac_sp_cereal, ac_sp_legume, bc_sp_cereal,bc_sp_legume, ac_sp, bc_sp)) %>% 
  # mutate(across(contains("biomass"), ~.x)) %>% #Pass biomass in g/m2
  mutate(across(contains("nitrogen_shoot"),~.x*0.1)) # Pass nitrogen_shoot in % instead of mg/g
                


#Table1
data_NNI <- data_compute_NNI_IC_SC %>%
  mutate(Nc_bedou_spi = p_spi*(ac_sp_i*biomass_shoot_total^(-bc_sp_i)) +
           (1-p_spi)*(ac_sp_j*biomass_shoot_total^(-bc_sp_j)),
         Nc_mix_spi = (ac_sp_i*biomass_shoot_total^(-bc_sp_i)),
         NNI_bedou = (nitrogen_shoot_component)/Nc_bedou_spi,
         NNI_mix = (nitrogen_shoot_component)/Nc_mix_spi)



data_NNI_IC <- data_NNI %>%
  filter(crop_type == "IC")

```


```{r NNI at flowering for comparison, eval = FALSE}
library(stringr)

date_flo <- traits_crop %>% 
  filter(experiment_id %in% c("Auzeville_cochard_2010","Auzeville_PP_2011","Auzeville_SGs_2007","Auzeville_TE_2006", "Auzeville_TO_2013", "Auzeville_ZN_2012"),
         str_detect(variable, "BBCH_65"),
         species_mix %in% c("wheat_turgidum", "pea", "fababean", "wheat_turgidum_pea", "wheat_turgidum_fababean")) %>%
  distinct(experiment_id, management, variable, species,  date_flo = date) 

data_NNI_flo <- data_NNI_IC %>%
  left_join(date_flo) %>%
  filter(date == date_flo) %>% 
  distinct(experiment_id, management,date_flo, plant_family, species, NNI_flo =  NNI_bedou, biomass_flo =  biomass_shoot_component, nitrogen_flo =  nitrogen_shoot_component)

data_NNI_harvest <- data_NNI_IC%>%
  left_join(index_paper %>% distinct(experiment_id, harvest_date, management)) %>%
  filter(date == harvest_date) %>% 
  distinct(experiment_id, management, date, plant_family, species, NNI_harvest = NNI_bedou, biomass_harvest =  biomass_shoot_component,  nitrogen_harvest = nitrogen_shoot_component)


data_compare_NNI_harvest_flo <- data_NNI_flo %>% inner_join(data_NNI_harvest) %>% 
  filter(!is.na(NNI_harvest), !is.na(NNI_flo))

data_compare_NNI_harvest_flo %>% 
  ggplot(aes(x = NNI_flo, y=  NNI_harvest)) + 
  geom_point() +
  facet_wrap(species~. ) +
  geom_abline(slope =1, intercept = 0, linetype = "dashed")

```

```{r plot INN flo vs harvest phd defence, eval = FALSE}
 

data_compare_NNI_harvest_flo <- data_NNI_flo %>% inner_join(data_NNI_harvest) %>% 
  filter(!is.na(NNI_harvest), !is.na(NNI_flo))

(plot_compare_NNI_harvest_flo <- data_compare_NNI_harvest_flo %>% 
  mutate(species =str_replace_all(species, c("fababean" = "Féverole", "pea" = "Pois", "wheat_turgidum" = "Blé dur"))) %>% 
  ggplot(aes(x = NNI_flo, y=  NNI_harvest)) + 
  geom_point() +
  facet_wrap(species~. ) +
  labs( y=  "INN à la récolte", x= "INN à la floraison") + 
  geom_abline(slope =1, intercept = 0, linetype = "dashed")+ 
  theme(axis.title = element_text(size = 18),
        axis.text = element_text(size=  18),
        strip.text = element_text(size= 17)))


(plot_compare_biomass_nitrogen_harvest_flo <- data_compare_NNI_harvest_flo %>% 
    select(experiment_id, management, plant_family, species, contains("biomass"), contains("nitrogen")) %>% 
    distinct() %>% 
    mutate(species =str_replace_all(species, c("fababean" = "Féverole", "pea" = "Pois", "wheat_turgidum" = "Blé dur"))) %>% 
    pivot_longer(matches("biomass|nitrogen"), names_sep = "_", names_to = c("variable", "moment")) %>% 
    pivot_wider(names_from = "moment") %>% 
    # filter(variable == "biomass") %>% 
    ggplot(aes(x = flo , y=  harvest)) + 
    geom_point(aes(color = experiment_id)) +
    # ggh4x::facet_grid2(species~variable, independent = "y", scales = "free_y") + 
    facet_grid(species~variable) + 
    labs( y=  "Valeur à la récolte", x= "Valeur à la floraison") + 
    geom_abline(slope =1, intercept = 0, linetype = "dashed")+ 
    theme(axis.title = element_text(size = 18),
          axis.text = element_text(size=  18),
          strip.text = element_text(size= 17)))


ggsave(plot = plot_compare_biomass_nitrogen_harvest_flo, filename = "C:/Users/rmahmoud/Documents/INRA/These/Soutenance thèse/figures_soutenance/plot_compare_biomass_nitrogen_harvest_flo.png", dpi = "retina", height =15, width  = 28, unit = "cm")

```


```{r plot NNI, eval = FALSE}

data_NNI_IC %>% 
  filter(!is.na(NNI_bedou)) %>%
  mutate(species_mix = stringr::str_replace_all(species_mix, c("wheat_turgidum" = "wt"))) %>% 
  ggplot(aes(x = species_mix, y =  NNI_bedou)) + 
  geom_point(position = position_jitterdodge(jitter.width = .2, jitter.height = 0), alpha = .7, aes(color = species)) + 
  geom_boxplot(alpha = .2, outlier.alpha = 0)+
  facet_wrap(plant_family ~., scales = "free_x" ) +
  geom_hline(linetype = "dashed", yintercept =  1) +
  scale_y_continuous(n.breaks = 6) +
  theme(axis.text.x = element_text(angle = -45, vjust = 0.5))


```



# Missing data


## Visualize

```{r}

data_viz_NA_NNI_IC <- data_NNI_IC %>% left_join(index_paper %>% distinct(experiment_id, harvest_date)) %>% select(experiment_id, species_mix, harvest_date, plant_family, date, contains("biomass"), contains("nitrogen_shoot"), contains("NNI"))

# vis_miss(data_viz_NA_NNI_IC)

# Only 2 missing values at harvest dates
gg_miss_upset(data_viz_NA_NNI_IC %>% filter(date == harvest_date) %>% select(contains("NNI_bedou"), contains("nitrogen"), contains('biomass')))


data_viz_NA_NNI_IC %>% filter(date == harvest_date, is.na(NNI_bedou))
# 

data_compare_dates <-data_NNI_IC %>%
  left_join(index_paper %>% distinct(experiment_id, harvest_date, management)) %>%
  select(experiment_id, species_mix, harvest_date, plant_family, 
         date, contains("biomass"), contains("nitrogen_shoot"),
         contains("NNI"), management) %>% 
  arrange(experiment_id, management,plant_family, date) %>%
  filter(!is.na(NNI_bedou)) %>%
  group_by(experiment_id, management, plant_family) %>% 
  summarise(last_NNI = NNI_bedou[which.max(date)],date_last_NNI = date[which(NNI_bedou == last_NNI)], last_date = max(date), harvest_date = unique(harvest_date), .groups = "drop" )

data_compare_dates %>% filter(date_last_NNI != harvest_date) # Ok last date is always harvest date


```

## Impute

```{r}
data_write_NNI_IC <-data_NNI_IC %>%
  left_join(index_paper %>% distinct(experiment_id, management,harvest_date)) %>%
  select(experiment_id, species_mix, harvest_date, plant_family, 
         date, contains("biomass"), contains("nitrogen_shoot"),
         contains("NNI"), management) %>% 
  arrange(experiment_id, management,plant_family, date) %>%
  filter(!is.na(NNI_bedou)) %>%
  group_by(experiment_id, management, plant_family) %>% 
  summarise(NNI_last = NNI_bedou[which.max(date)], .groups = "drop" ) %>% 
  pivot_wider(names_from = plant_family, values_from = NNI_last, names_prefix = "NNI_")


```


```{r impute 10 times the 2 missing values}

# data_write_NNI_IC %>% 
#   left_join(index_paper %>% 
#               distinct(experiment_id, management, species_mix)) %>% 
#   ggplot(aes(y = NNI_cereal, x = NNI_legume)) +
#   geom_point(aes(color = experiment_id), size = 3)  +
#   facet_wrap(species_mix~. )



data_impute_NNI <- data_write_NNI_IC %>%
  left_join(index_paper %>% distinct(experiment_id, management, species_mix)) %>%
  mutate(species_mix = as.factor(species_mix),
         experiment_id_integer = as.integer(as.factor(experiment_id))) %>% #change types of variables to set mice
  as.data.frame()


library(mice)
library(miceadds)
pred <- make.predictorMatrix(data_impute_NNI)
pred[, c( "management", "experiment_id")] <- 0
pred[c("NNI_cereal"), "experiment_id_integer"] <- -2

# PMM with hierarchical structure
mids <- mice(data = data_impute_NNI, predictorMatrix = pred, method = "2l.pmm", m = 10, printFlag = FALSE)
# mids <- mice(data = data_impute_NNI, predictorMatrix = pred, method = "norm", m = 10, printFlag = FALSE)

# check consistency 

# mids %>%
#   complete(action = "long") %>%
#   as_tibble() %>%
#   left_join(data_write_NNI_IC %>%
#               rename(NNI_cereal_raw = NNI_cereal) ) %>%
#   mutate(imputed= is.na(NNI_cereal_raw)) %>%
#   ggplot(aes(x = NNI_legume, y = NNI_cereal))  +
#   geom_point(aes(color = as.factor(experiment_id), shape = imputed)) +  facet_wrap(species_mix~., scales= "free")


data_write_NNI_imputed <-mids%>%
  complete(action = "long") %>%
  as_tibble() %>%
  select(-c(.id, experiment_id_integer)) 


```


# Write data

```{r write data, eval  = FALSE}

write_rds(path = "data/data_NNI.rds", x = data_write_NNI_imputed, compress = "xz")

```
