==> R/2-modeling/4-Impute_values_and_compute_differences.Rmd <==
---
title: "Impute values and compute differences between IC-SC and intra IC"
output:
  pdf_document: default
  word_document: default
  html_document: default
---

```{r, include = FALSE, message=F}
library(knitr)
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  echo = F,
  warning = F,
  message = F
)

```

```{r}

library(dplyr)
library(stringr)
library(tidyr)
library(purrr)
library(readr)
library(ggplot2)
theme_set(theme_bw())

# For missing data handling
library(naniar)
library(mice)
library(JointAI)

# library(sjmisc)

conflicted::conflict_prefer("select", "dplyr")
conflicted::conflict_prefer("filter", "dplyr")


```


# Set data

```{r data_global}
# load trait database and index
data_traits_mean <- data.intercrop::traits_mean
index <- data.intercrop::index %>% mutate(crop_id = paste(management, year, plot_name, crop_type, species, cultivar, sep = "_")) 

data_management_control_SC <- readr::read_rds("data/management_control_IC_SC.rds") %>% 
  mutate(management_control = ifelse(plant_family == "legume",
                                     management_SC_N0_control,
                                     management_SC_control)) 



data_asymp_wide <- read_rds("data/data_asymp.rds")%>% 
  select(-time_thermal_asymp) %>% 
  pivot_wider(values_from = asymp, names_from = variable, names_prefix = "asymp_" )

data_biomass_fits <- read_rds("data/data_biomass_fit.rds") %>%
  rename_with(.fn = ~ paste0(.,"_biomass"), .cols = asymp:lambda) # Add biomass column suffix at columns

data_height_fits <- read_rds("data/data_height_fit.rds")%>%
  rename_with(.fn = ~ paste0(.,"_height"), .cols = asymp:lambda) # Add height column suffix at columns

data_fits <- full_join(data_height_fits, data_biomass_fits) %>% 
  select(experiment_id, crop_id, matches("lambda|^slope")) %>%
  left_join(data_asymp_wide %>%
              select(-asymp_cover)) %>%
  left_join(index %>% distinct(experiment_id, crop_id, management, plant_family)) %>% 
  select(-crop_id) %>% 
  pivot_wider( names_from = plant_family, values_from = matches("height|biomass"))

data_SLA_LAI_cover <- read_rds("data/data_SLA_LAI_cover.rds")


data_all_predictors <- data_SLA_LAI_cover %>% 
  left_join(data_fits)


data_fits_tidy <- data_fits  %>% 
  pivot_longer(matches("height|biomass"), names_sep =  "(?=cereal|legume)", names_to = c("name", "plant_family"), values_drop_na = TRUE, values_to = "value_SC") %>%
  rename(management_control = management) %>% 
  mutate(name= str_remove(name, "_$"))

```



# Functions imputation

```{r custom missing data function}

# Re-write functions of naniar to visualize NA according to factors

custom_coerce_fct_na_explicit <- function(x){
  if (is.factor(x) & anyNA(x)) {
    forcats::fct_explicit_na(x, na_level = "NA")
  } else {
    x
  }
}
custom_gg_miss_fct <- function(x, fct){
  
  fct <- rlang::enquo(fct)
  
  data <- x %>%
    # protect against error where grouping by missing value leads to
    # warning message from dplyr about explicit
    dplyr::mutate_at(vars(!!fct), .funs = custom_coerce_fct_na_explicit) %>%
    dplyr::group_by(!!fct) %>%
    miss_var_summary()
  
  ggobject <-
    ggplot(data,
           aes_string(x = quo_name(fct),
                      y = "variable",
                      fill = "pct_miss")) +
    geom_tile() +
    viridis::scale_fill_viridis(name = "% de données\nmanquantes") +
    theme_minimal() +
    theme(axis.text.x = element_text(angle = 45,
                                     hjust = 1))
  
  return(ggobject)
}


source('R/scripts_functions/functions_tools.R')

```



# Visualize missing data

```{r}

data_all_predictors %>% 
  select(-c(experiment_id, management)) %>% 
  vis_miss()

custom_gg_miss_fct(data_all_predictors, experiment_id)
custom_gg_miss_fct(data_all_predictors %>% left_join(index %>% distinct(experiment_id, management, species_mix)), species_mix)
md.pattern(data_all_predictors %>% select(-c(experiment_id:cover_before_800_normalized_tt_SC_legume, contains("asymp"))))
```


```{r}

library(multilevel)


# High level of intra class correlation
data_all_predictors %>% pivot_longer(-c(experiment_id, management)) %>% 
  group_by( name) %>% 
  nest() %>% 
  ungroup %>% 
  mutate(icc = purrr::map_dbl(data, .f = function(x) ICC1(aov(x$value~x$experiment_id, data = x))))
```

```{r}


data_impute <- data_all_predictors %>% 
  select(-cover_before_800_normalized_tt, -cover_before_800_normalized_tt_SC_cereal,  -cover_before_800_normalized_tt_SC_legume) %>% 
  pivot_longer(-c(experiment_id, management)) %>% 
  separate(name, into = c("variable", "plant_family"), sep = "(?=cereal|legume)") %>% 
  mutate(variable = str_remove(variable, "_$")) %>% 
  pivot_wider(names_from = c(variable), values_from = "value") %>% 
  left_join(index %>% distinct(experiment_id, management, plant_family, species)) %>% 
  select(-c( plant_family))

# Suppress the values of lambda_height and lambda_biomass overestimated for these managements
data_impute[which(data_impute$experiment_id == "Auzeville_SGs_2007" & data_impute$management == "M35"), c("lambda_height", "slope_height")] <- NA
data_impute[which(data_impute$experiment_id == "Auzeville_cochard_2010" & data_impute$management == "M27"), c("lambda_biomass", "slope_biomass")] <- NA
data_impute[which(data_impute$experiment_id == "Auzeville_ZN_2012" & data_impute$management == "M27"), c("lambda_biomass", "slope_biomass")] <- NA

# M27_2010_cochard_IC_fababean_castel
# 
# data_impute %>%
#   select(-plant_family) %>%
#   pivot_longer(-c(experiment_id, management, species)) %>%
#   group_by( name) %>%
#   nest() %>%
#   ungroup %>%
#   mutate(icc = purrr::map_dbl(data, .f = function(x) ICC1(aov(x$value~x$experiment_id + x$species , data = x)))) %>%
#   ggplot(aes(x = name, y = icc)) +
#   geom_point() + facet_wrap(species~.)
custom_gg_miss_fct(data_impute %>% shorten_exp_id() , experiment_id) +
  

```


```{r figures imputation, eval = FALSE}

# plot_missing_by_experiment_id <- data_impute %>%select(-c(species, management, cover_before_800_normalized_tt_SC, asymp_biomass)) %>%  custom_gg_miss_fct(fct =experiment_id)


ggsave(filename = "figures/plot_missing_data_by_experiment_id.png", plot = plot_missing_by_experiment_id + labs(y  = NULL, x = NULL), width = 14, height = 18, unit ="cm")



vis_miss(data_impute %>% select(-c(experiment_id, management, species)))

library(latex2exp)


plot_missing_data_by_xp_id <- data_impute %>% shorten_exp_id() %>% select(-c(species, management, asymp_biomass)) %>% 
  select(experiment_id, asymp_height, slope_height, lambda_height, slope_biomass, lambda_biomass, contains("max_SLA"), contains("max_LAI")) %>%   custom_gg_miss_fct(experiment_id)+
  scale_y_discrete(labels=c('slope_biomass'=parse(text = TeX('$ \\mu_{biom}$')),
                            "max_LAI_C" = parse(text = TeX("$, LAI, C$")),
                            "max_LAI_L" = parse(text = TeX("$, LAI, L$")),
                            'lambda_height'=parse(text = TeX('$\\lambda_{height}$')),
                            'lambda_biomass'=parse(text = TeX('$\\lambda_{biom}$')),
                            'max_sla_GLT_SC' = parse(text = TeX('$max(SLA)_{SC}$')), 
                            'max_LAI' = parse(text = TeX('$max(LAI)$')), 
                            'max_LAI_SC' = parse(text = TeX('$max(LAI)_{SC}$')), 
                            'max_sla_GLT' = parse(text = TeX('$max, SLA$')), 
                            'slope_height'=parse(text = TeX('$ \\mu_{height}$')),
                            'asymp_height'=parse(text = TeX('$max(height)$')),
                            "density_factor" = "density")) + 
  labs(x = NULL, y= NULL) +
  theme(legend.text = element_text(size= 8), legend.title =element_text(size = 8))

ggsave(filename = "figures/plot_missing_by_experiment_id.png", plot = plot_missing_data_by_xp_id + labs(y  = NULL, x = NULL), width = 14, height = 10, unit ="cm")
  file.copy("figures/plot_missing_by_experiment_id.png", to  ="C:/Users/rmahmoud/Dropbox/Applications/Overleaf/Manuscrit thèse/chap4_modeling_IC/figures/plot_missing_by_experiment_id.png", overwrite = TRUE)


data_count_NA <- data_impute %>% 
  mutate(species = str_replace_all(species, c("fababean" = "legume", "pea"= "legume"))) %>% 
  pivot_wider(names_from = species, values_from = where(is.numeric))

apply(data_count_NA, 1, function(x)any(is.na(x))) %>% sum
```

```{r plot missing data soutenance, eval = FALSE}
plot_missing_data_by_xp_id_soutenance <- data_impute %>%
  shorten_exp_id() %>% 
  filter(experiment_id %in% c("Auz_2013", "Auz_cochard_2010", "Auz_SGs_2007")) %>% 
  select(-c(species, management, asymp_biomass)) %>% 
  select(experiment_id, slope_height, slope_biomass,  max_LAI) %>%   custom_gg_miss_fct(experiment_id)+
  scale_y_discrete(labels=c('slope_biomass'=parse(text = TeX('$ \\mu_{biom}$')),
                            "max_LAI_C" = parse(text = TeX("$, LAI, C$")),
                            "max_LAI_L" = parse(text = TeX("$, LAI, L$")),
                            'lambda_height'=parse(text = TeX('$\\lambda_{height}$')),
                            'lambda_biomass'=parse(text = TeX('$\\lambda_{biom}$')),
                            'max_sla_GLT_SC' = parse(text = TeX('$max(SLA)_{SC}$')), 
                            'max_LAI' = parse(text = TeX('$max(LAI)$')), 
                            'max_LAI_SC' = parse(text = TeX('$max(LAI)_{SC}$')), 
                            'max_sla_GLT' = parse(text = TeX('$max, SLA$')), 
                            'slope_height'=parse(text = TeX('$ \\mu_{height}$')),
                            'asymp_height'=parse(text = TeX('$max(height)$')),
                            "density_factor" = "density")) + 
  labs(x = NULL, y= NULL) +
  theme(legend.text = element_text(size= 10), legend.title =element_text(size = 12), axis.text = element_text(size = 12))


ggsave(filename = "C:/Users/rmahmoud/Documents/INRA/These/Soutenance thèse/figures_soutenance/plot_missing_by_experiment_id.png", plot = plot_missing_data_by_xp_id_soutenance , width = 14, height = 10, unit ="cm")

```

# Impute parameters


```{r, eval = TRUE}

# a bit long to run

data_jointAI <- data_impute %>%  as.data.frame() %>% mutate(species=  as.factor(species))
plot_all(data_jointAI %>% mutate(experiment_id = as.factor(experiment_id)), idvars = c("experiment_id"))

#perform imputation
mod.jointAI <- lme_imp(asymp_biomass ~  max_LAI + max_LAI_SC + max_sla_GLT_SC + max_sla_GLT + slope_height + lambda_height + slope_biomass + lambda_biomass   + asymp_height + species + (1|experiment_id),
                       n.iter = 2000,
                       data = data_jointAI,
                       monitor_params = c(imps = TRUE),
                       trunc = list(max_LAI = c(1e-5, NA),
                                    max_LAI_SC = c(1e-5, NA),
                                    slope_height = c(1e-8, NA),
                                    slope_biomass = c(1e-8, NA)))

mod.jointAI <- lme_imp(asymp_biomass ~  max_LAI + max_LAI_SC + max_sla_GLT_SC + max_sla_GLT + slope_height + lambda_height + slope_biomass + lambda_biomass   + asymp_height + species + (asymp_height + 1|experiment_id),
                       n.iter = 2000,
                       data = data_jointAI,
                       monitor_params = c(imps = TRUE),
                       trunc = list(max_LAI = c(1e-5, NA),
                                    max_LAI_SC = c(1e-5, NA),
                                    slope_height = c(1e-8, NA),
                                    slope_biomass = c(1e-8, NA)))

# data_jointAI$experiment_id <- as.factor(data_jointAI$experiment_id)
# mod.jointAI <- lm_imp(asymp_biomass ~   species+max_LAI + max_LAI_SC + max_sla_GLT_SC + max_sla_GLT + slope_height + lambda_height + slope_biomass + lambda_biomass + asymp_height + experiment_id, 
#                        n.iter = 2000, 
#                        data = data_jointAI, 
#                        monitor_params = c(imps = TRUE),
#                        trunc = list(max_LAI = c(1e-5, NA),
#                                     max_LAI_SC = c(1e-5, NA),
#                                     slope_height = c(1e-8, NA),
#                                     slope_biomass = c(1e-8, NA)))


mod.jointAI$comp_info

readr::write_rds("data/model_imputation/mod.jointAI.rds",  x=  mod.jointAI, compress = "xz")




data_fits_impute <- data_impute%>%
  mutate(plant_family = ifelse(str_detect(species, "wheat"), "cereal", "legume")) %>%   select(experiment_id:management, plant_family, matches("biomass|height")) %>% 
  left_join(data_management_control_SC %>% distinct(experiment_id, management, plant_family, management_control)) %>% 
  pivot_longer(matches("height|biomass")) %>% 
  left_join(data_fits_tidy ) %>% 
  group_by( experiment_id, management, plant_family, name) %>% 
  summarise(value_SC =  value_SC,
            .groups = "drop") %>% 
  pivot_wider(names_from = c(name, plant_family), values_from = c(value_SC), names_glue = "{name}_{plant_family}_SC") %>% as.data.frame()
  
# data_fits_impute
data_fits_impute[which(data_fits_impute$experiment_id == "Auzeville_SGs_2007" & data_fits_impute$management %in% c("M39", "M47") ), c("slope_biomass_cereal_SC", "lambda_biomass_cereal_SC")] <- NA
data_fits_impute[which(data_fits_impute$experiment_id == "Auzeville_SGs_2007" & data_fits_impute$management %in% c("M20", "M46") ), c("slope_height_cereal_SC", "lambda_height_cereal_SC")] <- NA

mod.jointAI_SC_biomass <- lme_imp(asymp_biomass_cereal_SC ~  slope_biomass_cereal_SC + lambda_biomass_cereal_SC + asymp_height_cereal_SC + slope_height_cereal_SC + lambda_height_cereal_SC +  (1|experiment_id),
                       n.iter = 2000,
                       data = data_fits_impute,
                       monitor_params = c(imps = TRUE),
                       trunc = list(slope_biomass_cereal_SC = c(1e-8, NA)))

# data_fits_impute$experiment_id <- as.factor(data_fits_impute$experiment_id)
# mod.jointAI_SC_biomass <- lm_imp(asymp_biomass_cereal_SC ~  slope_biomass_cereal_SC + lambda_biomass_cereal_SC + asymp_height_cereal_SC + slope_height_cereal_SC + lambda_height_cereal_SC +  experiment_id, 
#                        n.iter = 2000, 
#                        data = data_fits_impute, 
#                        monitor_params = c(imps = TRUE),
#                        trunc = list(slope_biomass_cereal_SC = c(1e-8, NA)))

# traceplot(mod.jointAI_SC_biomass)
readr::write_rds("data/model_imputation/mod.jointAI_SC_biomass.rds",  x=  mod.jointAI_SC_biomass, compress = "xz")


```


```{r}

set.seed(1)
mod.jointAI <- readr::read_rds("data/model_imputation/mod.jointAI.rds")
summary(mod.jointAI, missinfo = TRUE)
# traceplots draw sausages, good point
traceplot(mod.jointAI)

impDF <- get_MIdat(mod.jointAI, m = 10, include = FALSE) %>% 
  as_tibble() %>% 
  rename(.imp = Imputation_)

mod.jointAI_SC_biomass <- readr::read_rds("data/model_imputation/mod.jointAI_SC_biomass.rds")
impDF_SC_biomass <- get_MIdat(mod.jointAI_SC_biomass, m = 10, include = FALSE) %>% 
  as_tibble() %>% 
  rename(.imp = Imputation_)
```




# Compute differences 

```{r}


data_diff_intra_IC <- impDF %>%
  select(.imp:management, !contains("SC") & !contains("rownr")) %>%
  mutate(plant_family = ifelse(str_detect(species, "wheat"), "cereal", "legume")) %>% 
  select(-c(species)) %>% 
  pivot_longer(-c(.imp, experiment_id, management, plant_family)) %>% 
  group_by(experiment_id, management, name, .imp) %>% 
  summarise(cereal = value[which(plant_family == "cereal")],
            legume = value[which(plant_family == "legume")],
            diff=  cereal - legume, .groups = "drop") %>% 
  pivot_wider(names_from = name, values_from = c(legume, cereal, diff)) %>% 
  rename_with(.fn = ~ paste0(str_remove(., "legume_"), "_legume"), contains("legume_"))%>% 
  rename_with(.fn = ~ paste0(str_remove(., "biomass_"), "_biomass"), contains("biomass_"))



data_imputation_SC_biomass_tidy <- impDF_SC_biomass  %>% 
  pivot_longer(matches("height|biomass"), names_sep =  "(?=cereal|legume)", names_to = c("name", "plant_family"), values_drop_na = TRUE, values_to = "value_SC") %>%
  mutate(name= str_remove(name, "_$")) %>% 
  mutate(plant_family = str_remove(plant_family, "_SC")) %>% 
  select(-.rownr)



data_diff_IC_SC_height_biomass <- impDF%>%
  mutate(plant_family = ifelse(str_detect(species, "wheat"), "cereal", "legume")) %>%   select(.imp:management, plant_family, matches("biomass|height")) %>% 
  left_join(data_management_control_SC %>% distinct(experiment_id, management, plant_family, management_control)) %>% 
  pivot_longer(matches("height|biomass")) %>% 
  left_join(data_imputation_SC_biomass_tidy ) %>% 
  left_join(index %>% distinct(experiment_id, management, plant_family, density_relative, density_factor)) %>%
  separate(density_factor, into = c("density_cereal", "density_legume"), sep = "_") %>%  
  group_by(.imp, experiment_id, management, plant_family, name) %>% 
  # summarise(scaling_factor = ifelse(name %in% c("max_LAI", "slope_biomass", "asymp_biomass"),
  #                                   as.numeric(unique(density_relative)), 1),
            summarise(scaling_factor = ifelse(name %in% c("slope_biomass", "asymp_biomass"),
                             as.numeric(density_relative)/(as.numeric(unique(density_cereal)) + as.numeric(unique(density_legume))), 1),
            diff_IC_SC = (value/scaling_factor)- value_SC,
            .groups = "drop") %>% 
  select(-scaling_factor) %>% 
  pivot_wider(names_from = c(name, plant_family), values_from = c(diff_IC_SC), names_prefix = "diff_IC_SC_")
  

data_diff_IC_SC_SLA_LAI <- impDF %>%
  mutate(plant_family = ifelse(str_detect(species, "wheat"), "cereal", "legume")) %>% 
  select(.imp:management, matches ("LAI|SLA"), plant_family) %>% 
  rename_with(matches("max_SLA|LAI|biomass|height") & !matches("SC"),.fn =  ~paste0(., '_IC')) %>%
  left_join(index %>% distinct(experiment_id, management, plant_family, density_relative, density_factor)) %>% 
  # left_join(data_fits_tidy) %>% 
  pivot_longer(-c(.imp, experiment_id, management, plant_family, density_relative, density_factor), names_sep = "_(?=SC|IC)", names_to=  c("name", "crop_type")) %>% 
  separate(density_factor, into = c("density_cereal", "density_legume"), sep = "_") %>% 
  group_by(.imp, experiment_id, management, plant_family, name) %>% 
  # summarise(
  #   scaling_factor = ifelse(name %in% c("max_LAI", "slope_biomass", "asymp_biomass"),
  #                                   as.numeric(unique(density_relative)), 1),
  summarise(
    scaling_factor = ifelse(name %in% c("slope_biomass", "asymp_biomass"),
                                    as.numeric(density_relative)/(as.numeric(unique(density_cereal)) + as.numeric(unique(density_legume))), 1),
  # summarise(
  #   scaling_factor = ifelse(name %in% c("slope_biomass", "asymp_biomass"),
  #                                   as.numeric(unique(density_relative)), 1),
            diff_IC_SC = value[which(crop_type == "IC")]/scaling_factor - value[which(crop_type == "SC")],
            .groups = "drop") %>% 
  distinct%>% 
  select(-c(scaling_factor)) %>% 
  pivot_wider(names_from = c(name, plant_family), values_from = c(diff_IC_SC), names_prefix = "diff_IC_SC_")

data_diff_IC_SC_cover <- data_SLA_LAI_cover %>% distinct(experiment_id, management, across(contains("cover")))

data_all_diff_with_imputations <- data_diff_IC_SC_height_biomass %>% 
  left_join(data_diff_intra_IC) %>% 
  left_join(data_diff_IC_SC_SLA_LAI) %>%
  left_join(data_diff_IC_SC_cover)



data_look_outliers <- data_all_diff_with_imputations %>% left_join(index %>% distinct(experiment_id, management, species_mix, N_amount)) 


```

# Write data





```{r}

write_rds(x = data_all_diff_with_imputations, "data/data_all_diff_with_imputations.rds")

```




# Look for outliers



```{r}

data_BE_CE_SE <- read_rds("data/data_BE_CE_SE_LER.rds")

data_look_outliers %>% 
  select(experiment_id, management, .imp, contains("diff"), contains("NNI"), species_mix, N_amount) %>% 
  select(-c(diff_IC_SC_asymp_biomass_legume, diff_IC_SC_asymp_biomass_cereal, diff_asymp_biomass)) %>% 
  pivot_longer(-c(experiment_id, management, .imp, species_mix, N_amount)) %>% 
  group_by(experiment_id, management, name, species_mix, N_amount) %>% 
  summarise(value = mean(value), .groups = "drop") %>% 
  left_join(data_BE_CE_SE %>% distinct(experiment_id, management, PLER_cereal)) %>% 
  shorten_exp_id() %>% 
  ggplot(aes(x = value, y=  PLER_cereal)) + 
  geom_point(aes(color = experiment_id, size =N_amount), alpha = 0.7 ) + 
      scale_size_continuous(range = c(1,3)) + 
      labs(x = NULL, y = "PLER céréale", color = "Expérimentation", size = "N (kg/ha)")  +
      facet_wrap(name~., scales= "free") +
      scale_colour_manual(values = c("Auz_2013" = "#F8766D",
                                    "Auz_cochard_2010" = "#B79F00",
                                    "Auz_PP_2011" = "#00BA38",
                                    "Auz_SGs_2007" = "#00BFC4",
                                    "Auz_TE_2006" = "#619CFF",
                                    "Auz_ZN_2012" = "#F564E3"
)) +
  theme(legend.position = "bottom") + 
      guides(colour = guide_legend(nrow = 2), size = guide_legend(nrow = 2), shape = guide_legend(nrow = 2) ) 


# 
data_look_outliers %>%
  select(experiment_id, management, .imp, contains("diff"), contains("NNI"), species_mix, N_amount) %>%
  select(-c(diff_IC_SC_asymp_biomass_legume, diff_IC_SC_asymp_biomass_cereal, diff_asymp_biomass)) %>%
  pivot_longer(-c(experiment_id, management, .imp, species_mix, N_amount)) %>%
  filter(str_detect(name, "LAI")) %>%
  group_by(experiment_id, management, name, species_mix, N_amount) %>%
  summarise(value = mean(value), .groups = "drop") %>%
  left_join(data_BE_CE_SE %>% distinct(experiment_id, management, PLER_cereal)) %>%
  shorten_exp_id() %>%
  ggplot(aes(x = value, y=  PLER_cereal)) +
  geom_point(aes(color = experiment_id, size =N_amount), alpha = 0.7 ) +
      scale_size_continuous(range = c(3,5)) +
      labs(x = NULL, y = "PLER céréale", color = "Expérimentation", size = "N (kg/ha)")  +
      facet_wrap(name~., scales= "free") +
      scale_colour_manual(values = c("Auz_2013" = "#F8766D",
                                    "Auz_cochard_2010" = "#B79F00",
                                    "Auz_PP_2011" = "#00BA38",
                                    "Auz_SGs_2007" = "#00BFC4",
                                    "Auz_TE_2006" = "#619CFF",
                                    "Auz_ZN_2012" = "#F564E3"
)) +
  theme(legend.position = "bottom") +
      guides(colour = guide_legend(nrow = 2), size = guide_legend(nrow = 2), shape = guide_legend(nrow = 2) )


```

```{r}

library(latex2exp)
# data_look_outliers <- data_all_diff_with_imputations %>% left_join(index %>% distinct(experiment_id, management, species_mix)) 

data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_pea") %>% 
  select(experiment_id, management, .imp, species_mix, diff_lambda_height, lambda_height_legume, cereal_lambda_height,  diff_slope_biomass, slope_legume_biomass, cereal_slope_biomass , diff_lambda_biomass,lambda_legume_biomass, cereal_lambda_biomass,diff_slope_height, slope_height_legume, cereal_slope_height )%>% 
  pivot_longer(-c(experiment_id, species_mix, management, .imp)) %>% 
  mutate(name = forcats::fct_inorder(name)) %>% 
  group_by(experiment_id, management, name) %>% 
  mutate(imputed = ifelse(value[1] == value[2], FALSE, TRUE)) %>% 
  ungroup %>% 
  ggplot(aes(x = name, y = value)) + 
  geom_boxplot(outlier.size =0) +
  facet_wrap(name~., scales ="free", ncol =3) +
  geom_point( position = position_jitterdodge(jitter.width = .01), aes(color = experiment_id, shape = imputed)) +
  # theme(legend.position = "bottom") +
  labs(x = NULL, y = NULL)


data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_pea", slope_legume_biomass > 0.02)  %>% 
  select(diff_slope_biomass, slope_legume_biomass, cereal_slope_biomass, diff_IC_SC_slope_biomass_legume, diff_lambda_height, .imp, management)


data_look_outliers %>% 
  # filter(species_mix == "wheat_turgidum_fababean", diff_IC_SC_lambda_height_cereal < - 190)  %>%
  filter(species_mix == "wheat_turgidum_fababean",management == "M36", experiment_id == "Auzeville_ZN_2012")  %>% 
  select(diff_lambda_height,diff_IC_SC_lambda_height_cereal, lambda_height_legume, cereal_lambda_height, diff_IC_SC_lambda_height_cereal, diff_lambda_height, .imp, management)

data_look_outliers %>% 
  # filter(species_mix == "wheat_turgidum_fababean", diff_IC_SC_lambda_height_cereal < - 190)  %>%
  filter(species_mix == "wheat_turgidum_fababean",management == "M36", experiment_id == "Auzeville_ZN_2012")  %>% 
  select(diff_lambda_height, diff_IC_SC_asymp_height_cereal,  diff_IC_SC_slope_height_cereal, .imp, management)



# data_look_outliers %>% filter(diff_IC_SC_lambda_biomass_cereal < - 200)  %>% 
#   select(diff_lambda_biomass,diff_IC_SC_lambda_biomass_cereal, diff_IC_SC_lambda_biomass_legume, lambda_legume_biomass, cereal_lambda_biomass, diff_IC_SC_lambda_biomass_cereal, diff_lambda_biomass, .imp, management, experiment_id, species_mix) %>% arrange(management, species_mix, diff_lambda_biomass) %>% View



data_look_outliers %>% 
  # filter(species_mix == "wheat_turgidum_fababean", diff_IC_SC_lambda_height_cereal < - 190)  %>%
  filter(species_mix == "wheat_turgidum_fababean", diff_IC_SC_lambda_height_cereal < -100)  %>% 
  select(diff_lambda_height,diff_IC_SC_lambda_height_cereal, lambda_height_legume, cereal_lambda_height, diff_IC_SC_lambda_height_cereal, diff_lambda_height, .imp, management, experiment_id, species_mix)


# data_look_outliers %>% 
#   # filter(species_mix == "wheat_turgidum_fababean", diff_IC_SC_lambda_height_cereal < - 190)  %>%
#   filter(species_mix == "wheat_turgidum_fababean",management == "M36", experiment_id == "Auzeville_ZN_2012")  %>% 
#   select(diff_lambda_height, diff_IC_SC_asymp_height_cereal,  diff_IC_SC_slope_height_cereal, .imp, management)





data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_pea", diff_lambda_height < -400)  %>% 
  select(experiment_id , management, diff_lambda_height, lambda_height_legume, cereal_lambda_height, slope_height_legume, cereal_slope_height)


data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_pea",lambda_height_legume > 750)  %>% 
  select(experiment_id , management, diff_lambda_height, lambda_height_legume, cereal_lambda_height, slope_height_legume, cereal_slope_height)



data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_pea", diff_slope_height >0.0010)  %>% 
  distinct(experiment_id , management, .imp,slope_height_legume, cereal_slope_height, diff_slope_height) %>% count(experiment_id, management)


data_look_outliers %>% 
  filter(diff_IC_SC_lambda_height_cereal > 200) %>%  
  select(experiment_id , management, diff_IC_SC_lambda_height_cereal, diff_lambda_height, lambda_height_legume, cereal_lambda_height) %>% 
  mutate(cereal_lambda_height_SC = cereal_lambda_height - diff_IC_SC_lambda_height_cereal)

```




```{r}

data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean") %>% 
  select(experiment_id, management, .imp, species_mix, diff_lambda_height, lambda_height_legume, cereal_lambda_height,  diff_slope_biomass, slope_legume_biomass, cereal_slope_biomass , diff_lambda_biomass,lambda_legume_biomass, cereal_lambda_biomass,diff_slope_height, slope_height_legume, cereal_slope_height )%>% 
  pivot_longer(-c(experiment_id, species_mix, management, .imp)) %>% 
  mutate(name = forcats::fct_inorder(name)) %>% 
  ggplot(aes(x = name, y = value, fill = name)) + 
  geom_boxplot(outlier.size =0) +
  facet_wrap(name~., scales ="free", ncol =3) +
  geom_abline(intercept = 0, linetype = "dashed", slope = 0) + 
  geom_point(pch = 21, position = position_jitterdodge(jitter.width = .01)) +
  theme(legend.position = "none") + 
  labs(x = NULL, y = NULL)



data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean") %>% 
  select(experiment_id, management, .imp, species_mix, diff_lambda_height, lambda_height_legume, cereal_lambda_height,  diff_slope_biomass, slope_legume_biomass, cereal_slope_biomass , diff_lambda_biomass,lambda_legume_biomass, cereal_lambda_biomass,diff_slope_height, slope_height_legume, cereal_slope_height )%>% 
  pivot_longer(-c(experiment_id, species_mix, management, .imp)) %>% 
  mutate(name = forcats::fct_inorder(name)) %>% 
  group_by(experiment_id, management, name) %>% 
  mutate(imputed = ifelse(value[1] == value[2], FALSE, TRUE)) %>% 
  ungroup %>% 
  ggplot(aes(x = name, y = value)) + 
  geom_boxplot(outlier.size =0) +
  facet_wrap(name~., scales ="free", ncol =3) +
  geom_point( position = position_jitterdodge(jitter.width = .01), aes(color = experiment_id, shape = imputed)) +
  # theme(legend.position = "bottom") +
  labs(x = NULL, y = NULL)


```



```{r}
data_look_outliers %>% 
  filter(diff_IC_SC_lambda_height_cereal > 200) %>% select(diff_IC_SC_lambda_height_cereal , cereal_lambda_height , experiment_id, management, .imp, species_mix) %>% 
  mutate(cereal_lambda_height_SC = cereal_lambda_height - diff_IC_SC_lambda_height_cereal) 


data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean", diff_lambda_biomass > 450)  %>% 
  select(experiment_id , management, species_mix, diff_lambda_biomass, lambda_legume_biomass, cereal_lambda_biomass, slope_legume_biomass, cereal_slope_biomass)

data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean", diff_lambda_height >200)  %>% 
  select(experiment_id , management, diff_lambda_height, lambda_height_legume, cereal_lambda_height, slope_height_legume, cereal_slope_height)


data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean", diff_lambda_biomass < -300)  %>% 
  select(experiment_id , management, .imp, diff_lambda_biomass, lambda_legume_biomass, cereal_lambda_biomass, slope_legume_biomass, cereal_slope_biomass)


data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean", cereal_slope_height > 0.0025)  %>% 
  select(experiment_id , management, diff_lambda_height, lambda_height_legume, cereal_lambda_height, slope_height_legume, cereal_slope_height)


data_look_outliers %>% 
  filter(species_mix == "wheat_turgidum_fababean", slope_height_legume < 0.00075)  %>% 
  select(experiment_id , management, diff_lambda_height, lambda_height_legume, cereal_lambda_height, slope_height_legume, cereal_slope_height)




```

```{r}

data.intercrop::traits_mean %>% 
  filter(variable == "biomass_shoot", experiment_id == "Auzeville_cochard_2010", management == "M27")  %>% 
  distinct(variable, species, management, time_thermal, time_sowing, value) %>% 
  ggplot(aes(x = time_thermal, y=  value)) + 
  geom_point(aes(color = species), position = position_jitter(width = 0, height = .01)) + 
  geom_smooth(aes(group = species), method ="lm", se =FALSE)

data.intercrop::traits_mean %>% 
  filter(variable == "biomass_shoot", experiment_id == "Auzeville_TO_2013", management == "M60")  %>% 
  distinct(variable, species, management, time_thermal, time_sowing, value) %>% 
  ggplot(aes(x = time_thermal, y=  value)) + 
  geom_point(aes(color = species), position = position_jitter(width = 0, height = .01)) + 
  geom_smooth(aes(group = species), method ="lm", se =FALSE)


data.intercrop::traits_mean %>% 
  filter(variable == "height", experiment_id == "Auzeville_TO_2013", management == "M32")  %>% 
  distinct(variable, species, management, time_thermal, time_sowing, value) %>% 
  ggplot(aes(x = time_thermal, y=  value)) + 
  geom_point(aes(color = species), position = position_jitter(width = 0, height = .01)) + 
  geom_smooth(aes(group = species), method ="lm", se =FALSE)


data.intercrop::traits_mean %>% 
  filter( experiment_id == "Auzeville_ZN_2012", management == "M36", variable == "biomass_shoot")  %>% 
  distinct(variable, species, management, time_thermal, time_sowing, value) %>% 
  ggplot(aes(x = time_thermal, y=  value)) + 
  geom_point(aes(color = species), position = position_jitter(width = 0, height = .01)) + 
  geom_smooth(aes(group = species), method ="lm", se =FALSE)


```




```{r}

# Valeurs bizarres Auz2012 --> val cereal_slope height surestimées ?
data_look_outliers %>% select(experiment_id, management, .imp, cereal_slope_height) %>% filter(experiment_id == "Auzeville_ZN_2012") %>% arrange(management)

# Valeurs bizarres Auz2013 --> val cereal_slope height surestimées ?
data_look_outliers %>% select(experiment_id, management, .imp, diff_slope_biomass, slope_legume_biomass, cereal_slope_biomass) %>% filter(experiment_id == "Auzeville_TO_2013", diff_slope_biomass < -0.009) %>% arrange(management)

data_look_outliers %>% select(experiment_id, management, .imp, diff_slope_biomass, slope_legume_biomass, cereal_slope_biomass) %>% filter(experiment_id == "Auzeville_TO_2013", management == "M41") %>% arrange(management)



# Valeur bizarre --> slopê_height legume surestimée ?
data_look_outliers %>% select(experiment_id, management, .imp, diff_IC_SC_slope_height_legume, slope_height_legume) %>% filter(experiment_id == "Auzeville_TO_2013", diff_IC_SC_slope_height_legume < -6e-04)%>% arrange(management)



data_look_outliers %>% select(experiment_id, management,species_mix,  .imp, diff_IC_SC_max_sla_GLT_cereal, cereal_max_sla_GLT) %>% filter(experiment_id == "Auzeville_ZN_2012", diff_IC_SC_max_sla_GLT_cereal < -80)%>% arrange(management)
data_management_control_SC %>% filter(experiment_id == "Auzeville_ZN_2012", plant_family == "cereal", management %in% c("M23", "M24", "M26", "M32", "M33", "M34", "M35"))

data_look_outliers%>% 
  select(experiment_id, management, species_mix, .imp, diff_IC_SC_max_sla_GLT_cereal, cereal_max_sla_GLT )  %>% 
  filter(experiment_id == "Auzeville_ZN_2012") %>% 
  mutate(cereal_max_sla_GLT_SC = cereal_max_sla_GLT - diff_IC_SC_max_sla_GLT_cereal)  %>% 
  distinct(cereal_max_sla_GLT_SC)
  


data_look_outliers %>%
  select(experiment_id, management, species_mix, .imp, diff_IC_SC_max_sla_GLT_cereal, cereal_max_sla_GLT )  %>%
  mutate(cereal_max_sla_GLT_SC = cereal_max_sla_GLT - diff_IC_SC_max_sla_GLT_cereal) %>%
  group_by(experiment_id, management, species_mix) %>%
  summarise(diff_IC_SC_max_sla_GLT_cereal = mean(diff_IC_SC_max_sla_GLT_cereal), cereal_max_sla_GLT= mean(cereal_max_sla_GLT), cereal_max_sla_GLT_SC = mean(cereal_max_sla_GLT_SC), .groups = "drop") %>%
  pivot_longer( - c(diff_IC_SC_max_sla_GLT_cereal,experiment_id, management, species_mix)) %>%
  mutate(wrongSLA = ifelse(diff_IC_SC_max_sla_GLT_cereal < -50, TRUE, FALSE) ) %>%
  ggplot(aes(x = value, diff_IC_SC_max_sla_GLT_cereal))  +
  geom_point(aes(color = experiment_id, shape =wrongSLA), size = 2) +
  facet_grid(species_mix~name ) +
  # facet_wrap(species_mix~. ) +
  theme(legend.position = "bottom")


data_look_outliers %>% 
  select(experiment_id, management,species_mix,  .imp, diff_max_LAI, max_LAI_legume, cereal_max_LAI) %>% filter(experiment_id == "Auzeville_SGs_2007")%>% arrange(management) %>% 
  group_by(management) %>% 
  mutate(imputed = ifelse(diff_max_LAI[1] != diff_max_LAI[2], TRUE, FALSE)) %>% 
  ungroup %>%
  pivot_longer(-c(experiment_id, species_mix, management, .imp, imputed)) %>% 
  ggplot(aes(x = name, y = value)) + 
  geom_boxplot(outlier.size =0) +
  # facet_wrap(name~., scales ="free", ncol =3) +
  geom_abline(intercept = 0, linetype = "dashed", slope = 0) + 
  geom_point(pch = 21, aes(color = imputed), position = position_jitter(width = .02, height = 0)) +
  labs(x = NULL, y = NULL)


data_look_outliers %>%
  # select(.imp, experiment_id, management, contains("diff"), species_mix) %>% 
  pivot_longer(-c(experiment_id, species_mix, management, .imp)) %>% 
  filter(name %in% c("diff_IC_SC_max_sla_GLT_cereal", "cereal_max_sla_GLT")) %>% 
  group_by(experiment_id, management, name) %>% 
  summarise(value = mean(value), .groups = "drop") %>% 
  mutate(name = forcats::fct_inorder(name)) %>% 
  ggplot(aes(x = name, y = value)) + 
  geom_boxplot(outlier.size =0) +
  facet_wrap(name~., scales ="free", ncol =3) +
  geom_abline(intercept = 0, linetype = "dashed", slope = 0) + 
  geom_point(pch = 21, aes(color = experiment_id), position = position_jitter(width = .02)) +
  theme(legend.position = "none") + 
  labs(x = NULL, y = NULL)

data_look_outliers %>%
  # select(.imp, experiment_id, management, contains("diff"), species_mix) %>% 
  pivot_longer(-c(experiment_id, species_mix, management, .imp)) %>% 
  filter(name %in% c("diff_IC_SC_max_sla_GLT_cereal", "diff_IC_SC_max_sla_GLT_legume", "diff_max_LAI", "diff_IC_SC_slope_height_legume", "diff_slope_height", "slope_height_legume", "cereal_slope_height")) %>% 
  group_by(experiment_id, management, name) %>% 
  summarise(value = mean(value), .groups = "drop") %>% 
  mutate(name = forcats::fct_inorder(name)) %>% 
  ggplot(aes(x = name, y = value)) + 
  geom_boxplot(outlier.size =0) +
  facet_wrap(name~., scales ="free", ncol =3) +
  geom_abline(intercept = 0, linetype = "dashed", slope = 0) + 
  geom_point(pch = 21, aes(color = experiment_id), position = position_jitter(width = .02)) +
  theme(legend.position = "none") + 
  labs(x = NULL, y = NULL)




data_look_outliers %>%
  # select(.imp, experiment_id, management, contains("diff"), species_mix) %>% 
  mutate(cereal_slope_height_SC = cereal_slope_height - diff_IC_SC_slope_height_cereal) %>%
  mutate(cereal_lambda_height_SC = cereal_lambda_height - diff_IC_SC_lambda_height_cereal) %>%
  mutate(cereal_asymp_height_SC = cereal_asymp_height - diff_IC_SC_asymp_height_cereal) %>%
  pivot_longer(-c(experiment_id, species_mix, management, .imp)) %>% 
  filter(name %in% c("diff_IC_SC_slope_height_cereal","cereal_slope_height", "cereal_slope_height_SC", "diff_IC_SC_lambda_height_cereal",  "cereal_lambda_height" ,"cereal_lambda_height_SC", "diff_IC_SC_asymp_height_cereal", "cereal_asymp_height", "cereal_asymp_height_SC")) %>% 
  group_by(experiment_id, management, species_mix, name) %>%
  summarise(value = mean(value), .groups = "drop") %>%
  group_by(experiment_id, management, species_mix) %>% 
  mutate(diff_IC_SC = case_when(
    str_extract(name, "asymp|lambda|slope") == "asymp" ~ value[which(name =="diff_IC_SC_asymp_height_cereal")],
    str_extract(name, "asymp|lambda|slope") == "lambda" ~ value[which(name =="diff_IC_SC_lambda_height_cereal")],
    str_extract(name, "asymp|lambda|slope") == "slope" ~ value[which(name =="diff_IC_SC_slope_height_cereal")],
    )) %>% 
  filter(!str_detect(name, "diff_IC_SC")) %>% 
  mutate(name = forcats::fct_inorder(name)) %>% 
  ggplot(aes(x = value, y = diff_IC_SC)) + 
  facet_wrap(name~., scales ="free", ncol =2) +
  geom_abline(intercept = 0, linetype = "dashed", slope = 0) + 
  geom_point(pch = 21, aes(color = experiment_id), position = position_jitter(width = .02)) +
  theme(legend.position = "none") + 
  labs(x = NULL, y = NULL)



data_look_outliers %>% 
  filter(diff_IC_SC_asymp_height_cereal > 0.2) %>% 
  select(experiment_id, management, .imp, diff_IC_SC_asymp_height_cereal, cereal_asymp_height) %>% 
  mutate(cereal_asymp_height_SC = cereal_asymp_height - diff_IC_SC_asymp_height_cereal)



data_look_outliers %>%
  filter(diff_IC_SC_asymp_height_cereal < -0.05) %>% 
  select(experiment_id, management, .imp, diff_IC_SC_asymp_height_cereal, cereal_asymp_height) %>% 
  mutate(cereal_asymp_height_SC = cereal_asymp_height - diff_IC_SC_asymp_height_cereal)



# max_LAI + max_LAI_SC + max_sla_GLT_SC + max_sla_GLT + slope_height + lambda_height + slope_biomass + lambda_biomass   + asymp_height + species 
data_look_outliers %>%
  filter(experiment_id == "Auzeville_ZN_2012") %>% 
  select(experiment_id, management, .imp, diff_IC_SC_asymp_height_cereal, cereal_asymp_height, cereal_max_LAI, cereal_max_sla_GLT , cereal_lambda_biomass, cereal_slope_biomass) %>% 
  mutate(cereal_asymp_height_SC = cereal_asymp_height - diff_IC_SC_asymp_height_cereal) %>% 
  ggplot(aes(x = cereal_max_sla_GLT, cereal_asymp_height, shape = ifelse(management == "M36", TRUE, FALSE))) + 
    geom_point() 


data_look_outliers%>% 
  distinct(experiment_id, management, .imp, diff_max_LAI, diff_IC_SC_max_LAI_legume, max_LAI_legume)  %>%
  mutate(max_LAI_legume_SC = max_LAI_legume - diff_IC_SC_max_LAI_legume)%>% 
  arrange(experiment_id, management) %>% 
  group_by(experiment_id, management) %>% 
  summarise(max_LAI_legume_IC = mean(max_LAI_legume), max_LAI_legume_SC = mean(max_LAI_legume_SC), .groups = "drop") %>% 
  left_join(index %>% distinct(experiment_id, species_mix, management, density_factor))  %>% 
  ggplot(aes(x = max_LAI_legume_IC, max_LAI_legume_SC)) + 
  geom_point(aes(shape = density_factor, color = experiment_id)) + 
  facet_wrap(species_mix ~. , scales = "free") + 
  geom_y_x() + 
  theme(legend.position = "bottom")

data_look_outliers %>% filter(diff_IC_SC_max_LAI_legume < -4) %>% distinct(experiment_id, management, .imp, diff_max_LAI, diff_IC_SC_max_LAI_legume, max_LAI_legume)  %>%
  mutate(max_LAI_legume_SC = max_LAI_legume - diff_IC_SC_max_LAI_legume)%>% arrange(experiment_id, management) %>% View

# data_look_outliers %>% filter(diff_IC_SC_max_LAI_cereal > 4) %>% distinct(experiment_id, management, .imp, diff_max_LAI, diff_IC_SC_max_LAI_cereal, cereal_max_LAI)  %>% 
#   mutate(cereal_max_LAI_SC = cereal_max_LAI/0.5 - diff_IC_SC_max_LAI_cereal)%>% arrange(experiment_id, management) %>% View

  
```
