==> R/2-modeling/5-Run_all_models.Rmd <==
---
title: "Run all models MERF LongituRF"
output:
  pdf_document: default
  word_document: default
  html_document: default
---

```{r, include = FALSE, message=F}
library(knitr)
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  echo = F,
  warning = F,
  message = F
)

```

```{r}

library(dplyr)
library(stringr)
library(tidyr)
library(purrr)
library(readr)
library(ggplot2)


# Modeling
library(nlme)
library(randomForest)
library(Boruta)

theme_set(theme_bw())

# library(sjmisc)

conflicted::conflict_prefer("select", "dplyr")
conflicted::conflict_prefer("filter", "dplyr")


```


```{r functions}
# Read all functions
list_functions <- list.files("R/scripts_functions", full.names = TRUE)
invisible(lapply(list_functions, source))

#wrapper around gmerf

# fun_call_MERF_modified <- function(data, y = "yield_cereal", group = "experiment_id", max_iterations = 200, toll_param = 0.02, verbose_param = FALSE, list_control_lme = list(msMaxIter = 1000, msMaxEval = 1000, opt = "optim"), weights_lme = varIdent(form = ~1|group) ){
#   return(gmerf_modified(y = data[[y]],  cov = data %>% select(-any_of(y), - management, -any_of(group)), group = as.factor(data[[group]]), family = "gaussian", itmax = max_iterations, weights = weights_lme, control_lme = list_control_lme, toll = toll_param,verbose = verbose_param)) 
# }


fun_call_MERF_modified <- function(data, y = "yield_cereal", group = "experiment_id", max_iterations = 200, toll_param = 0.001, verbose_param = FALSE ){
  
   return(MERF_modified(X =  as.data.frame(data %>% select(-any_of(y), - management, -any_of(group))), Y =  data[[y]], id = data[[group]], Z = as.matrix(rep(1, nrow(data))), time = rep(1,nrow(data)),  iter = max_iterations, verbose = verbose_param, delta= toll_param))
}


nest_data_by_imp <- function(data){return(data %>% group_by(.imp) %>% nest() %>% ungroup)}

```


```{r}
data_all_diff <-  read_rds( "data/data_all_diff_with_imputations.rds")

data_BE_CE_SE <- read_rds("data/data_BE_CE_SE_LER.rds")

data_NNI <- read_rds("data/data_NNI.rds") %>%
  inner_join(data_all_diff %>% distinct(experiment_id, management)) %>% 
  shorten_exp_id()

data_diff_wt_fababean <- data_all_diff%>%
  left_join(data.intercrop::index %>% distinct(experiment_id, management, species_mix)) %>%
  filter(species_mix== "wheat_turgidum_fababean") %>%
  shorten_exp_id()


index <- data.intercrop::index %>% 
  shorten_exp_id()%>%
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management) ) 


```


# Document structure

For each model, the yield performance variable is extracted. Then, the model is fitted within specific chunks (as this may be long to run).


# Model yield cereal WT / fababean

## Yield

```{r}

itk_wt_fababean <- index %>% distinct(experiment_id, management, cultivar_mix, N_amount, mixing_pattern, mixture_design,  density_factor ) %>% 
  separate(cultivar_mix, into = c("cult_cereal", "cult_legume"), sep = "_")

```


```{r}

# Relevant features

yield_cereal_wt_fababean <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "cereal") %>%
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_cereal = value)


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "yield_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(yield_cereal_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_cereal_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, max_iterations =  3000)) 

```

```{r, eval = TRUE}

write_rds(models_yield_cereal_wt_fababean, "models/MERF/wt_fababean/yield_cereal_wt_fababean.rds")


```

## PLER


```{r}

# Relevant features

LER_wt_fababean <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, PLER_cereal) 


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "PLER_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(LER_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_PLER_cereal_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "PLER_cereal", max_iterations = 2000)) 

```

```{r, eval = TRUE}

write_rds(models_PLER_cereal_wt_fababean, "models/MERF/wt_fababean/PLER_cereal_wt_fababean.rds")


```

# Model yield legume WT / fababean


## Yield

```{r}

# Relevant features


yield_legume_wt_fababean <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "legume") %>%
  group_by(experiment_id, management) %>% 
  summarise(value = max(value), .groups = "drop") %>%  #get maximal value of the yield for management M36 as it is measured twice
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_legume = value)

covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "yield_legume")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(yield_legume_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_legume), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_legume_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "yield_legume", max_iterations =  2000)) 

```



```{r, eval = TRUE}

write_rds(models_yield_legume_wt_fababean, "models/MERF/wt_fababean/yield_legume_wt_fababean.rds")


```

## PLER




```{r}

# Relevant features

LER_wt_fababean <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, PLER_legume) 


covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "PLER_legume")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(LER_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_legume), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_PLER_legume_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "PLER_legume", max_iterations = 2000)) 


```

```{r, eval = TRUE}

write_rds(models_PLER_legume_wt_fababean, "models/MERF/wt_fababean/PLER_legume_wt_fababean.rds")


```

# Model yield total WT / fababean

## Yield

```{r}

# Relevant features


yield_total_wt_fababean <- yield_legume_wt_fababean %>%
  left_join(yield_cereal_wt_fababean) %>% 
  mutate(yield_total = yield_legume + yield_cereal) %>%
  distinct(experiment_id, management, yield_total) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "yield_total")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(yield_total_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_total_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "yield_total", max_iterations = 3000)) 

```



```{r, eval = TRUE}

write_rds(models_yield_total_wt_fababean, "models/MERF/wt_fababean/yield_total_wt_fababean.rds")


```


## LER




```{r}

# Relevant features

LER_wt_fababean <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, LER) 


covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "LER")) 

itk_covariates <- c("cult_cereal", "cult_legume")


data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(LER_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_LER_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "LER", max_iterations = 2000)) 

```

```{r, eval = TRUE}

write_rds(models_LER_wt_fababean, "models/MERF/wt_fababean/LER_wt_fababean.rds")


```

## Biodiversity effects WT / fababean


### BE

```{r}

# Relevant features


BE_wt_fababean <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, BE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "BE")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(BE_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_BE_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "BE", max_iterations = 3000)) 

```



```{r, eval = FALSE}

write_rds(models_BE_wt_fababean, "models/MERF/wt_fababean/BE_wt_fababean.rds")


```


### CE


```{r}

# Relevant features


CE_wt_fababean <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, CE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "CE")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(CE_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_CE_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "CE", max_iterations = 2000)) 

```


```{r, eval = FALSE}

write_rds(models_CE_wt_fababean, "models/MERF/wt_fababean/CE_wt_fababean.rds")


```




### SE


```{r}

# Relevant features


SE_wt_fababean <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, SE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "SE")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_fababean <- data_diff_wt_fababean %>% 
  left_join(SE_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_SE_wt_fababean <- data_wt_fababean_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "SE", max_iterations =  1000)) 

```


```{r, eval = FALSE}

write_rds(models_SE_wt_fababean, "models/MERF/wt_fababean/SE_wt_fababean.rds")


```



# Set data wt / pea



```{r}

data_diff_wt_pea <- data_all_diff%>%
  left_join(data.intercrop::index %>% distinct(experiment_id, management, species_mix)) %>%
  filter(species_mix== "wheat_turgidum_pea") %>%
  shorten_exp_id()


index_wt_pea <- data.intercrop::index %>% 
  shorten_exp_id()%>%
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management) ) 

itk_wt_pea <- index_wt_pea %>% distinct(experiment_id, management, cultivar_mix, N_amount, mixing_pattern, mixture_design,  density_factor ) %>% 
  separate(cultivar_mix, into = c("cult_cereal", "cult_legume"), sep = "_")

```


# Model yield cereal WT / pea

## Yield



```{r}

# Relevant features

yield_cereal_wt_pea <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "cereal") %>%
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_cereal = value)


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "yield_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(yield_cereal_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_cereal_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, max_iterations =  3000, toll_param = 0.0002)) 

```

```{r, eval = TRUE}

write_rds(models_yield_cereal_wt_pea, "models/MERF/wt_pea/yield_cereal_wt_pea.rds")


```

## PLER


```{r}

# Relevant features

PLER_wt_pea <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, PLER_cereal) 


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "PLER_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(PLER_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_PLER_cereal_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "PLER_cereal")) 

```

```{r, eval = TRUE}

write_rds(models_PLER_cereal_wt_pea, "models/MERF/wt_pea/PLER_cereal_wt_pea.rds")


```

# Model yield legume WT / pea


## Yield

```{r}

# Relevant features


yield_legume_wt_pea <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "legume") %>%
  group_by(experiment_id, management) %>% 
  summarise(value = max(value), .groups = "drop") %>%  #get maximal value of the yield for management M36 as it is measured twice
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_legume = value)

covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "yield_legume")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(yield_legume_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_legume), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_legume_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "yield_legume", max_iterations =  3000)) 

```



```{r, eval = TRUE}

write_rds(models_yield_legume_wt_pea, "models/MERF/wt_pea/yield_legume_wt_pea.rds")


```

## PLER


```{r}

# Relevant features

PLER_wt_pea <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, PLER_legume) 


covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "PLER_legume")

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(PLER_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_legume), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_PLER_legume_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "PLER_legume", max_iterations = 2000))  

```

```{r, eval = TRUE}

write_rds(models_PLER_legume_wt_pea, "models/MERF/wt_pea/PLER_legume_wt_pea.rds")


```

# Model yield total WT / pea

## Yield

```{r}

# Relevant features


yield_cereal_wt_pea <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "cereal") %>%
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_cereal = value)

yield_legume_wt_pea <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "legume") %>%
  group_by(experiment_id, management) %>% 
  summarise(value = max(value), .groups = "drop") %>%  #get maximal value of the yield for management M36 as it is measured twice
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_legume = value)


yield_total_wt_pea <- yield_legume_wt_pea %>%
  left_join(yield_cereal_wt_pea) %>% 
  mutate(yield_total = yield_legume + yield_cereal) %>%
  distinct(experiment_id, management, yield_total) 



covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "yield_cereal")


covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "yield_legume")


covariates_total <- unique(c( covariates_legume[!grepl("yield_legume|cover_before_800_normalized_tt_SC_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal|cover_before_800_normalized_tt_SC_cereal", covariates_cereal)],  "yield_total")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(yield_total_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_total_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "yield_total", max_iterations=  3000)) 

```



```{r, eval = TRUE}

write_rds(models_yield_total_wt_pea, "models/MERF/wt_pea/yield_total_wt_pea.rds")


```

## LER



```{r}

# Relevant features


LER_wt_pea <-  data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, LER) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "LER")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(LER_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_LER_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "LER", max_iterations=  1000)) 

```



```{r, eval = TRUE}

write_rds(models_LER_wt_pea, "models/MERF/wt_pea/LER_wt_pea.rds")


```



## Biodiversity effects WT / pea


### BE

```{r}

# Relevant features


BE_wt_pea <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, BE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "BE")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(BE_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_BE_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "BE", max_iterations = 2000)) 

```



```{r, eval = FALSE}

write_rds(models_BE_wt_pea, "models/MERF/wt_pea/BE_wt_pea.rds")


```


### CE


```{r}

# Relevant features


CE_wt_pea <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, CE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "CE")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(CE_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_CE_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "CE", max_iterations = 3000)) 

```


```{r, eval = FALSE}

write_rds(models_CE_wt_pea, "models/MERF/wt_pea/CE_wt_pea.rds")


```




### SE


```{r, eval = TRUE}

# Relevant features


SE_wt_pea <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, SE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "SE")) 

itk_covariates <- c("cult_cereal", "cult_legume")

data_wt_pea <- data_diff_wt_pea %>% 
  left_join(SE_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp

```


```{r, eval = FALSE}
models_SE_wt_pea <- data_wt_pea_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "SE", max_iterations = 2000))  


write_rds(models_SE_wt_pea, "models/MERF/wt_pea/SE_wt_pea.rds")


```






# Set data all species



```{r}

data_diff_all_species <- data_all_diff%>%
  left_join(data.intercrop::index %>% distinct(experiment_id, management, species_mix)) %>%
  shorten_exp_id()


index_all_species <- data.intercrop::index %>% 
  shorten_exp_id()%>%
  inner_join(data_diff_all_species %>% distinct(experiment_id, management) ) 

itk_all_species <- index_all_species %>% distinct(experiment_id, management, species_mix, cultivar_mix, N_amount, mixing_pattern, mixture_design,  density_factor ) %>% 
  separate(cultivar_mix, into = c("cult_cereal", "cult_legume"), sep = "_") %>% 
  mutate(legume = str_extract(species_mix, "pea|fababean")) %>% 
  select(-species_mix)

```


# Model yield cereal all species

## Yield



```{r}

# Relevant features

yield_cereal_all_species <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "cereal") %>%
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_cereal = value)


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "yield_cereal")

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(yield_cereal_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_cereal_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, max_iterations = 3000)) 

```

```{r, eval = TRUE}

write_rds(models_yield_cereal_all_species, "models/MERF/all_species/yield_cereal_all_species.rds")


```

## PLER


```{r}

# Relevant features

PLER_cereal_all_species <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, PLER_cereal) 


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "PLER_cereal")

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(PLER_cereal_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_PLER_cereal_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "PLER_cereal", max_iterations = 2000)) 

```

```{r, eval = TRUE}

write_rds(models_PLER_cereal_all_species, "models/MERF/all_species/PLER_cereal_all_species.rds")


```

# Model yield legume all species


## Yield

```{r}

# Relevant features


yield_legume_all_species <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "legume") %>%
  group_by(experiment_id, management) %>% 
  summarise(value = max(value), .groups = "drop") %>%  #get maximal value of the yield for management M36 as it is measured twice
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_legume = value)

covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "yield_legume")

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(yield_legume_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_legume), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_legume_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "yield_legume", max_iterations = 3000)) 

```



```{r, eval = TRUE}

write_rds(models_yield_legume_all_species, "models/MERF/all_species/yield_legume_all_species.rds")


```

## PLER


```{r}

# Relevant features

PLER_all_species <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, PLER_legume) 


covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "PLER_legume")

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(PLER_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  select(.imp, experiment_id, management, any_of(covariates_legume), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp


```



```{r, eval = TRUE}
models_PLER_legume_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "PLER_legume", max_iterations = 2000, toll_param = 0.02))  

```

```{r, eval = TRUE}

write_rds(models_PLER_legume_all_species, "models/MERF/all_species/PLER_legume_all_species.rds")


```

# Model yield total all species

## Yield

```{r}

# Relevant features


yield_cereal_all_species <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "cereal") %>%
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_cereal = value)

yield_legume_all_species <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "biomass_seed", plant_family == "legume") %>%
  group_by(experiment_id, management) %>% 
  summarise(value = max(value), .groups = "drop") %>%  #get maximal value of the yield for management M36 as it is measured twice
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, yield_legume = value)


yield_total_all_species <- yield_legume_all_species %>%
  left_join(yield_cereal_all_species) %>% 
  mutate(yield_total = yield_legume + yield_cereal) %>%
  distinct(experiment_id, management, yield_total) 



covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "yield_cereal")


covariates_legume <- c( "diff_IC_SC_asymp_height_legume", "diff_IC_SC_lambda_biomass_legume", "diff_IC_SC_lambda_height_legume", "diff_IC_SC_slope_biomass_legume", "diff_IC_SC_slope_height_legume",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI",  "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_legume", "diff_IC_SC_max_sla_GLT_legume", "NNI_legume",  "yield_legume")


covariates_total <- unique(c( covariates_legume[!grepl("yield_legume|cover_before_800_normalized_tt_SC_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal|cover_before_800_normalized_tt_SC_cereal", covariates_cereal)],  "yield_total")) 

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(yield_total_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_yield_total_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "yield_total", max_iterations=  3000)) 

```



```{r, eval = TRUE}

write_rds(models_yield_total_all_species, "models/MERF/all_species/yield_total_all_species.rds")


```

## LER



```{r}

# Relevant features


LER_all_species <-  data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, LER) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "LER")) 

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(LER_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```



```{r, eval = TRUE}
models_LER_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "LER", max_iterations=  1000)) 

```



```{r, eval = TRUE}

write_rds(models_LER_all_species, "models/MERF/all_species/LER_all_species.rds")


```



## Biodiversity effects all species


### BE

```{r}

# Relevant features


BE_all_species <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, BE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "BE")) 

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(BE_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_BE_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "BE", max_iterations = 3000)) 

```



```{r, eval = FALSE}

write_rds(models_BE_all_species, "models/MERF/all_species/BE_all_species.rds")


```


### CE


```{r}

# Relevant features


CE_all_species <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, CE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "CE")) 

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(CE_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```



```{r, eval = FALSE}
models_CE_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "CE", max_iterations = 2000)) 

```


```{r, eval = FALSE}

write_rds(models_CE_all_species, "models/MERF/all_species/CE_all_species.rds")


```




### SE


```{r}

# Relevant features


SE_all_species <- data_BE_CE_SE %>%
  shorten_exp_id() %>% 
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, SE) 

covariates_total <- unique(c( covariates_legume[!grepl("yield_legume", covariates_legume)], covariates_cereal[!grepl("yield_cereal", covariates_cereal)],  "SE")) 

itk_covariates <- c("cult_cereal", "cult_legume", "legume")

data_all_species <- data_diff_all_species %>% 
  left_join(SE_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  mutate(across(matches("experiment_id|cult"), as.factor)) %>% 
  select(.imp, experiment_id, management, any_of(covariates_total), any_of(itk_covariates))

data_all_species_nested <- data_all_species %>% nest_data_by_imp

```


```{r, eval = FALSE}
models_SE_all_species <- data_all_species_nested %>% 
  mutate(mod.MERF =  map(data, fun_call_MERF_modified, verbose_param = TRUE, y = "SE", max_iterations = 2000))  


write_rds(models_SE_all_species, "models/MERF/all_species/SE_all_species.rds")


```
