==> R/2-modeling/5-Run_all_models_N_seed_CV.Rmd <==
---
title: "Run all models cross validation MERF LongituRF"
output:
  pdf_document: default
  word_document: default
  html_document: default
---

```{r, include = FALSE, message=F}
library(knitr)
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  echo = F,
  warning = F,
  message = F
)

```

```{r}

library(dplyr)
library(stringr)
library(tidyr)
library(purrr)
library(readr)
library(ggplot2)




library(rsample)


# Modeling
library(nlme)
library(randomForest)
library(Boruta)

theme_set(theme_bw())

# library(sjmisc)

conflicted::conflict_prefer("select", "dplyr")
conflicted::conflict_prefer("filter", "dplyr")


```


```{r functions}
# Read all functions
list_functions <- list.files("R/scripts_functions", full.names = TRUE)
invisible(lapply(list_functions, source))

#wrapper around gmerf
# 
# fun_call_MERF_modified <- function(data, y = "N_seed_cereal", group = "experiment_id", max_iterations = 200, toll_param = 0.02, verbose_param = FALSE, list_control_lme = list(msMaxIter = 1000, msMaxEval = 1000, opt = "optim"), weights_lme = varIdent(form = ~1|group) ){
#   return(gmerf_modified(y = data[[y]],  cov = data %>% select(-any_of(y), - management, -any_of(group)), group = as.factor(data[[group]]), family = "gaussian", itmax = max_iterations, weights = weights_lme, control_lme = list_control_lme, toll = toll_param,verbose = verbose_param)) 
# }


performance_variables <- c("N_seed_cereal", "N_seed_legume", "N_seed_total", "PLER_cereal", "PLER_legume", "LER", "BE", "CE", "SE")

fun_call_MERF_modified <- function(data, performance_metrics = performance_variables,  y = "N_seed_cereal", group = "experiment_id", max_iterations = 200, toll_param = 0.0005, verbose_param = FALSE ){
  
   return(MERF_modified(X =  as.data.frame(data %>% select(-any_of(performance_metrics), - management, -any_of(group))), Y =  data[[y]], id = data[[group]], Z = as.matrix(rep(1, nrow(data))), time = rep(1,nrow(data)),  iter = max_iterations, verbose = verbose_param, delta= toll_param))
}


nest_data_by_imp <- function(data){return(data %>% group_by(.imp) %>% nest() %>% ungroup)}



get_data_test <- function(data, id_fold, data_kfold = data_kfold_wt_fababean ){
  return(data %>% 
           inner_join(data_kfold %>% filter(id == id_fold)%>% select(-id)))
}

get_data_train <- function(data, id_fold, data_kfold = data_kfold_wt_fababean ){
  return(data %>% 
           inner_join(data_kfold %>% filter(id != id_fold) %>% select(-id)))
}


```


```{r}
data_all_diff <-  read_rds( "data/data_all_diff_with_imputations.rds")

data_BE_CE_SE <- read_rds("data/data_BE_CE_SE_LER.rds")

data_NNI <- read_rds("data/data_NNI.rds") %>%
  inner_join(data_all_diff %>% distinct(experiment_id, management)) %>% 
  shorten_exp_id()



```

# Set data wt / fababean

```{r}

data_diff_wt_fababean <- data_all_diff%>%
  left_join(data.intercrop::index %>% distinct(experiment_id, management, species_mix)) %>%
  filter(species_mix== "wheat_turgidum_fababean") %>%
  shorten_exp_id()


set.seed(2022)
data_kfold_wt_fababean <- data_diff_wt_fababean %>%
  distinct(experiment_id, management) %>%
  vfold_cv(v = 4 , strata = experiment_id) %>%
  mutate(testing = map(splits, testing)) %>%
  unnest(testing) %>% 
  distinct(id, experiment_id, management)



index <- data.intercrop::index %>% 
  shorten_exp_id()%>%
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management) ) 


```

# Model N_seed cereal WT / fababean

## N_seed / PLER

```{r}

itk_wt_fababean <- index %>% distinct(experiment_id, management, cultivar_mix, N_amount, mixing_pattern, mixture_design,  density_factor ) %>% 
  separate(cultivar_mix, into = c("cult_cereal", "cult_legume"), sep = "_")

# Relevant features

N_seed_cereal_wt_fababean <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "nitrogen_seed", plant_family == "cereal") %>%
  inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, N_seed_cereal = value)
# 
# LER_wt_fababean <- data_BE_CE_SE %>%
#   shorten_exp_id() %>% 
#   inner_join(data_diff_wt_fababean %>% distinct(experiment_id, management)) %>% 
#   distinct(experiment_id, management, PLER_cereal) 


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI", "diff_max_sla_GLT", "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "N_seed_cereal", "PLER_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")




data_wt_fababean <- data_diff_wt_fababean %>% 
  inner_join(N_seed_cereal_wt_fababean) %>% 
  left_join(itk_wt_fababean) %>% 
  left_join(data_NNI) %>% 
  # inner_join(subset_train_wt_fababean) %>%
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))


data_wt_fababean_nested <- data_wt_fababean %>% nest_data_by_imp



data_kv_wt_fababean <- data_wt_fababean_nested %>%
  left_join(expand(data = data_wt_fababean_nested, .imp, fold_test = paste0("Fold",1:4) ))%>% 
  mutate(fold_test = as.list(fold_test)) %>% 
  mutate(data_test = map2(data, fold_test, get_data_test ))%>% 
  mutate(data_train = map2(data, fold_test, get_data_train ))

```



```{r, eval = TRUE}
models_N_seed_cereal_wt_fababean <- data_kv_wt_fababean %>% 
  mutate(mod.MERF =  map(data_train, fun_call_MERF_modified, verbose_param = TRUE, max_iterations =  3000, toll_param = 0.001)) 


write_rds(models_N_seed_cereal_wt_fababean, "models/N_seed_CV/wt_fababean/N_seed_cereal_CV_wt_fababean.rds")


```



# Set data wt / pea

```{r}

data_diff_wt_pea <- data_all_diff%>%
  left_join(data.intercrop::index %>% distinct(experiment_id, management, species_mix)) %>%
  filter(species_mix== "wheat_turgidum_pea") %>%
  shorten_exp_id()


set.seed(2022)
data_kfold_wt_pea <- data_diff_wt_pea %>%
  distinct(experiment_id, management) %>%
  vfold_cv(v = 4 , strata = experiment_id) %>%
  mutate(testing = map(splits, testing)) %>%
  unnest(testing) %>% 
  distinct(id, experiment_id, management)



index <- data.intercrop::index %>% 
  shorten_exp_id()%>%
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management) ) 


```

# Model N_seed cereal WT / pea

## N_seed / PLER

```{r}

itk_wt_pea <- index %>% distinct(experiment_id, management, cultivar_mix, N_amount, mixing_pattern, mixture_design,  density_factor ) %>% 
  separate(cultivar_mix, into = c("cult_cereal", "cult_legume"), sep = "_")

# Relevant features

N_seed_cereal_wt_pea <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "nitrogen_seed", plant_family == "cereal") %>%
  inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, N_seed_cereal = value)

# LER_wt_pea <- data_BE_CE_SE %>%
#   shorten_exp_id() %>% 
#   inner_join(data_diff_wt_pea %>% distinct(experiment_id, management)) %>% 
#   distinct(experiment_id, management, PLER_cereal) 


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI", "diff_max_sla_GLT", "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "N_seed_cereal", "PLER_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")




data_wt_pea <- data_diff_wt_pea %>% 
  # left_join(LER_wt_pea) %>% 
  inner_join(N_seed_cereal_wt_pea) %>% 
  left_join(itk_wt_pea) %>% 
  left_join(data_NNI) %>% 
  # inner_join(subset_train_wt_pea) %>%
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))


data_wt_pea_nested <- data_wt_pea %>% nest_data_by_imp



data_kv_wt_pea <- data_wt_pea_nested %>%
  left_join(expand(data = data_wt_pea_nested, .imp, fold_test = paste0("Fold",1:4) ))%>% 
  mutate(fold_test = as.list(fold_test)) %>% 
  mutate(data_test = map2(data, fold_test, get_data_test, data_kfold_wt_pea ))%>% 
  mutate(data_train = map2(data, fold_test, get_data_train, data_kfold_wt_pea ))

```



```{r, eval = TRUE}
models_N_seed_cereal_wt_pea <- data_kv_wt_pea %>% 
  mutate(mod.MERF =  map(data_train, fun_call_MERF_modified, verbose_param = TRUE, max_iterations =  3000, toll_param = 0.001)) 

```

```{r, eval = TRUE}

write_rds(models_N_seed_cereal_wt_pea, "models/N_seed_CV/wt_pea/N_seed_cereal_CV_wt_pea.rds")


```


# Set data all species



```{r}

data_diff_all_species <- data_all_diff%>%
  left_join(data.intercrop::index %>% distinct(experiment_id, management, species_mix)) %>%
  shorten_exp_id()



set.seed(2022)
data_kfold_all_species <- data_diff_all_species %>%
  distinct(experiment_id, management) %>%
  vfold_cv(v = 4 , strata = experiment_id) %>%
  mutate(testing = map(splits, testing)) %>%
  unnest(testing) %>% 
  distinct(id, experiment_id, management)



index <- data.intercrop::index %>% 
  shorten_exp_id()%>%
  inner_join(data_diff_all_species %>% distinct(experiment_id, management) ) 



```

# Model N_seed cereal all species

## N_seed / PLER

```{r}

itk_all_species <- index %>% distinct(experiment_id, management, cultivar_mix, N_amount, mixing_pattern, mixture_design,  density_factor ) %>% 
  separate(cultivar_mix, into = c("cult_cereal", "cult_legume"), sep = "_")

# Relevant features

N_seed_cereal_all_species <- data.intercrop::traits_mean %>%
  shorten_exp_id() %>% 
  filter(variable == "nitrogen_seed", plant_family == "cereal") %>%
  inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
  distinct(experiment_id, management, N_seed_cereal = value)

# LER_all_species <- data_BE_CE_SE %>%
#   shorten_exp_id() %>% 
#   inner_join(data_diff_all_species %>% distinct(experiment_id, management)) %>% 
#   distinct(experiment_id, management, PLER_cereal) 


covariates_cereal <- c( "diff_IC_SC_asymp_height_cereal", "diff_IC_SC_lambda_biomass_cereal", "diff_IC_SC_lambda_height_cereal", "diff_IC_SC_slope_biomass_cereal", "diff_IC_SC_slope_height_cereal",  "diff_asymp_height","diff_lambda_biomass", "diff_lambda_height", "diff_max_LAI", "diff_max_sla_GLT", "diff_slope_biomass", "diff_slope_height", "diff_IC_SC_max_LAI_cereal", "diff_IC_SC_max_sla_GLT_cereal", "NNI_cereal",  "N_seed_cereal", "PLER_cereal")

itk_covariates <- c("cult_cereal", "cult_legume")




data_all_species <- data_diff_all_species %>% 
  # left_join(LER_all_species) %>% 
  inner_join(N_seed_cereal_all_species) %>% 
  left_join(itk_all_species) %>% 
  left_join(data_NNI) %>% 
  # inner_join(subset_train_all_species) %>%
  select(.imp, experiment_id, management, any_of(covariates_cereal), any_of(itk_covariates))


data_all_species_nested <- data_all_species %>% nest_data_by_imp



data_kv_all_species <- data_all_species_nested %>%
  left_join(expand(data = data_all_species_nested, .imp, fold_test = paste0("Fold",1:4) ))%>% 
  mutate(fold_test = as.list(fold_test)) %>% 
  mutate(data_test = map2(data, fold_test, get_data_test, data_kfold_all_species ))%>% 
  mutate(data_train = map2(data, fold_test, get_data_train, data_kfold_all_species ))

```



```{r, eval = TRUE}
models_N_seed_cereal_all_species <- data_kv_all_species %>% 
  mutate(mod.MERF =  map(data_train, fun_call_MERF_modified, verbose_param = TRUE, max_iterations =  3000, toll_param = 0.001)) 


write_rds(models_N_seed_cereal_all_species, "models/N_seed_CV/all_species/N_seed_cereal_CV_all_species.rds")


```
